Improve documentation, clean up code
