+++ package.json
@@ -18,7 +18,7 @@
     "nightwatch:docker": "nightwatch -c config/nightwatch.docker-compose.js --retries 3",
     "nightwatch": "nightwatch -c config/nightwatch.js",
     "nightwatch-sauce": "nightwatch -c config/nightwatch-sauce.js",
-    "nightwatch-visual": "node src/platform/testing/visual-regression-puppeteer/index.js",
+    "nightwatch-visual": "node src/platform/testing/visual-regression/index.js",
     "postinstall": "npm rebuild node-sass",
     "reset:env": "./script/reset-environment.sh",
     "test": "npm run test:unit && npm run test:e2e",
@@ -45,6 +45,8 @@
     "test:watch": "node script/watch.js",
     "test:visual": "./script/run-nightwatch.sh --visual",
     "test:visual:baseline": "./script/run-nightwatch.sh --visual --command=baseline",
+    "test:visual:mobile": "./script/run-nightwatch.sh --visual --mobile",
+    "test:visual:mobile:baseline": "./script/run-nightwatch.sh --visual --command=baseline --mobile",
     "watch": "node script/build.js --watch",
     "watch:brand-consolidation": "node script/build.js --brand-consolidation-enabled --watch",
     "watch:devpreview": "npm run watch:brand-consolidation",
