--- src/platform/testing/visual-regression-puppeteer/index.js
@@ -1,104 +1,0 @@
-const commandLineArgs = require('command-line-args');
-const Auth = require('../e2e-puppeteer/auth');
-const puppeteer = require('puppeteer');
-const devices = require('puppeteer/DeviceDescriptors');
-const Timeouts = require('../e2e-puppeteer/timeouts');
-const { sitemapURLs } = require('../../site-wide/tests/sitemap/sitemap-helpers');
-const createBaselineImage = require('./util/create-baseline-image');
-const calculateDiff = require('./util/calculate-diff');
-const chalk = require('chalk');
-
-const iPhone = devices['iPhone 6'];
-
-
-const commands = {
-  CREATE_BASELINE_IMAGES: 'baseline',
-  CALCULATE_DIFFS: 'diff',
-  MOBILE: true
-
-};
-
-const { command, mobile } = commandLineArgs([
-  { name: 'command', type: String },
-  { name: 'config', type: String, alias: 'c' },
-  { name: 'mobile', type: Boolean }
-]);
-
-// Converts the array of routes/URL's and returns a super-long promise chain.
-async function createRouteHandlerChain(page, routes, routeHandler) {
-
-  // Loops through all of the routes continually chaining promises onto a single original promise.
-  // This way the promises will execute in a waterfall effect and is necessary since we can only show one URL at a time.
-  for (let i = routes.length - 1; i >= 0; i--) {
-    const route = routes[i];
-    /* eslint-disable no-await-in-loop, no-console */
-    try {
-      await page.goto(route, {
-        timeout: 0
-      });
-      await page.waitFor('body');
-      await routeHandler(page, route);
-    } catch (e) {
-      console.error(e);
-      console.log(chalk.red(`Couldn't screenshot ${route}`));
-    }
-    /* eslint-enable */
-    // await page.goto('about:blank')
-  }
-}
-
-// A wrapper around the login helper to return a promise
-async function login(page) {
-  const token = Auth.getUserToken();
-  await Auth.logIn(token, page, '/', 1);
-  await page.waitForSelector('body', { timeout: Timeouts.normal });
-}
-
-// Checks command flags to determine the "routeHandler" function that will perform a task after the browser navigates to each route.
-function getRouteHandler() {
-  switch (command) {
-    case commands.CREATE_BASELINE_IMAGES:
-      return createBaselineImage;
-    case commands.CALCULATE_DIFFS:
-    default:
-      return calculateDiff;
-  }
-}
-
-// Returns a function that will be executed as a Nightwatch test case.
-async function beginApplication(browser) {
-  const routeHandler = getRouteHandler();
-  // Before starting any routes, log the user in to prevent load errors.
-  const page = await browser.newPage();
-  if (mobile === commands.MOBILE) {
-    page.emulate(iPhone);
-  }
-  await login(page);
-  // Parse the sitemap XML file into an array of URL's
-  const sitemap = await sitemapURLs();
-  const routes = sitemap.urls;
-  await createRouteHandlerChain(page, routes, routeHandler);
-  await page.close();
-}
-
-// The entry point for everything.
-async function setup() {
-  // check if mobile test or not
-  const browser = await puppeteer.launch({ headless: false });
-  try {
-    await beginApplication(browser);
-  } catch (e) {
-  /* eslint-disable */
-    console.error(e, e.stack);
-  /* eslint-enable */
-  }
-  await browser.close();
-}
-
-try {
-  setup();
-} catch (e) {
-  /* eslint-disable */
-  console.log(e);
-
-}
