+++ src/platform/testing/visual-regression/util/calculate-diff.js
@@ -1,11 +1,14 @@
 const fs = require('fs');
 const resemble = require('node-resemble-js');
+const { expect }  = require('chai');
+const chalk = require('chalk');
+
 const { getFileNames, createDirectoryIfNotExist } = require('./get-file-names');
 
 const DIFF_THRESHOLD = 0.01;
 
 // A wrapper around fs.readFile to return a promise
-function readFile(fileName) {
+async function readFile(fileName) {
   return new Promise((resolve, reject) => {
     fs.readFile(fileName, (err, result) => {
       return err ? reject(err) : resolve(result);
@@ -15,15 +18,14 @@ function readFile(fileName) {
 
 // A wrapper around Nightwatch's screenshot function to return a promise
 // Note - Nightwatch's screenshot stores the screenshot only in memory as opposed to saveScreenshot
-function takeScreenshot(browser) {
-  const logScreenshotData = false;
-  return new Promise(resolve => browser.screenshot(logScreenshotData, resolve));
+async function takeScreenshot(browser) {
+  return browser.screenshot({ fullPage: true });
 }
 
 // Compares the values for the baseline image with the screenshot of the browser's current page.
-function executeComparison(baselineImageBuffer, screenshotResult) {
+async function executeComparison(baselineImageBuffer, screenshotResult) {
   const baselineImage = new Buffer(baselineImageBuffer, 'base64');
-  const screenshot = new Buffer(screenshotResult.value, 'base64');
+  const screenshot = new Buffer(screenshotResult, 'base64');
 
   // Execute ResembleJS to compare the images
   return new Promise(resolve => {
@@ -34,7 +36,7 @@ function executeComparison(baselineImageBuffer, screenshotResult) {
 }
 
 // Writes the diff image file used in the event the baseline doesn't match the screenshot of the browser's current page.
-function createDiffImage(diffFileName, comparisonResult) {
+async function createDiffImage(diffFileName, comparisonResult) {
   const diffImageStream = comparisonResult.getDiffImage();
   const writer = fs.createWriteStream(diffFileName);
 
@@ -53,7 +55,16 @@ function computeComparisonResult(browser, route, diffFileName, comparisonResult)
   const changesExceedThreshold = misMatchPercentage > DIFF_THRESHOLD;
 
   // Execution the test assertion
-  browser.verify.ok(!changesExceedThreshold, route);
+  try {
+    expect(changesExceedThreshold, `${route}`).to.be.false;
+  } catch (e) {
+    /* eslint-disable */
+    console.log(chalk.red(`${route} did not match up, diff is available at ${diffFileName} the mismatch was ${misMatchPercentage}`));
+    /* eslint-enable */
+
+  }
+
+  // browser.verify.ok(!changesExceedThreshold, route);
 
   // When the images differ, chain additional operations to create the diff image file
   if (changesExceedThreshold) {
@@ -70,45 +81,14 @@ function computeComparisonResult(browser, route, diffFileName, comparisonResult)
 }
 
 // The entry point for this module as a route handler
-function calculateDiff(browser, route) {
-  return new Promise((resolve, reject) => {
-    const timeout = 5000;
-    const [baselineFileName, diffFileName] = getFileNames(route);
-    const operations = [
-      readFile(baselineFileName),
-      takeScreenshot(browser)
-    ];
-
-    let isFinished = false;
-
-    // Set a timer just in case something weird happens and causes it to get stuck.
-    /* eslint-disable no-console */
-    setTimeout(() => {
-      if (!isFinished) {
-        console.log(`Operation timed out for the following route: ${route}`);
-        reject();
-      }
-    }, timeout);
-
-    Promise.all(operations)
-
-      // After reading both images into memory, run the operation used to compare their contents
-      .then(results => executeComparison(...results))
-
-      // Process the results from the comparison
-      .then(comparisonResult => computeComparisonResult(browser, route, diffFileName, comparisonResult))
-
-      // Calls to the wrapping function used for handling timeouts
-      .then(val => {
-        isFinished = true;
-        resolve(val);
-      })
-
-      .catch(err => {
-        isFinished = true;
-        reject(err);
-      });
-  });
+async function calculateDiff(browser, route) {
+  // const timeout = 5000;
+  const [baselineFileName, diffFileName] = getFileNames(route);
+  const fileBuffer = await readFile(baselineFileName);
+  const screenshotBuffer = await takeScreenshot(browser);
+  // After reading both images into memory, run the operation used to compare their contents
+  const comparisonResult = await executeComparison(fileBuffer, screenshotBuffer);
+  await computeComparisonResult(browser, route, diffFileName, comparisonResult);
 }
 
 module.exports = calculateDiff;
