+++ src/platform/testing/visual-regression/util/create-baseline-image.js
@@ -1,10 +1,11 @@
-const { getFileNames } = require('./get-file-names');
+const { getFileNames, createDirectoryIfNotExist } = require('./get-file-names');
 
 // The entry point for this module as a route handler.
 // Uses Nightwatch's native screenshot function to take a screenshot.
-function createBaselineImage(browser, route) {
+async function createBaselineImage(browser, route) {
   const [baselineFileName] = getFileNames(route);
-  return new Promise(resolve => browser.saveScreenshot(baselineFileName, resolve));
+  await createDirectoryIfNotExist(baselineFileName);
+  await browser.screenshot({ path: baselineFileName, fullPage: true, type: 'png' });
 }
 
 module.exports = createBaselineImage;
